Simplify calendar drill weekday calculation
